+++ src/main.rs
@@ -1,7 +1,9 @@
+extern crate processor;
 extern crate repl;
+extern crate table;
 
-use repl::*;
 use std::io::{self, Write};
+use table::Table;
 
 fn print_prompt() {
     print!("db > ");
@@ -19,16 +21,16 @@ fn main() {
             .expect("Failed to read line");
 
         if input_buffer.chars().next() == Some('.') {
-            match do_meta_command(&input_buffer, &table) {
+            match repl::do_meta_command(&input_buffer) {
                 Ok(_) => {}
                 Err(_) => {
                     println!("Unrecognized command {}", input_buffer);
                 }
             }
         } else {
-            match prepare_statement(&input_buffer) {
+            match processor::prepare_statement(&input_buffer) {
                 Ok(statement) => {
-                    execute_statement(statement, &mut table);
+                    processor::execute_statement(statement, &mut table);
                 }
                 Err(_) => {
                     println!("Unrecognized keyword at start of {}", input_buffer);
